perf: Factor out children handling to ensure re-renders as expected on scroll
